#9758: Refactored py files
