+++ ttnn/ttnn/__init__.py
@@ -322,4 +322,5 @@ from ttnn.operations.normalization import (
     determine_expected_group_norm_sharded_config_and_grid_size,
 )
 from ttnn.operations.conv2d import Conv2d, Conv2dConfig, get_conv_output_dim, get_conv_padded_input_shape_and_mem_config
-from ttnn.operations.pool import MaxPool2d, global_avg_pool2d, max_pool2d, average_pool_2d
+from ttnn.operations.max_pool2d import TTPyMaxPool, max_pool2d, max_pool2d_legacy, MaxPool2d
+from ttnn.operations.avg_pool2d import global_avg_pool2d, avg_pool2d
